Migrate pydsl tests to integration test suite
This marks the beginning of a migration of the pydsl renderer tests
to the integration test suite and away from the unit test suite.

The unit test approach is becoming too difficult to maintain due to its complex approach to trying to internally manage a highstate run instead of calling into the salt state system to fully manage it.
